Change project layout
- add cmd/trygolang/main.go
- add lib/mapping.go
